requirements.txt
fastapi
uvicorn
openai
pandas
spacy
python-dotenv
boto3  # For AWS Lambda
